Corner finding and masking
Fixed an issue with corner finding that was due to the corners being at the edge of the image, so I added a pad to the thresholded images. Added ability to mask edges of the input images. Increased rhombus threshold so that spherical distortion results in a switch to partial affine camera.