match confidence work around
It looks like OpenCV doesn't use the match confidence threshold passed in BestOf2NearestMatcher, so images that are too similar are given a confidence of zero and not matched. To get around this, I use the AffineBestOf2NearestMatcher instead when the confidence is too low. In the AffineBestOf2NearestMatcher source, the match threshold line is commented out so similar images are not rejected.