more straightening modifications
Previously, straightening a panorama would force the panorama to be split into at least two sections. Now if the panorama does not have multiple sections as dictated by the straighten_length, it is straightened as one image rather than being split. Added an option to save a low resolution version for easier viewing and sharing. Straighten length is still a parameter that should be changed based on how straight (and long) the batches are.