Updated straightening
Straightening is now more dynamic. To reduce warping, we look for points of interest for cutting the panorama. These are places in the midline with high second derivative in the direction normal to the camera movement. There are still some hardcoded thresholds for curvature and the spline point resolution, but it seems to work across different resolutions as is. For areas where points of interest are far apart (like when straightening the final panorama), it is necessary to add filler points so that each warped section is relatively small to stay within the OpenCV resolution limit for warpPerspective. Now does basic type checking in config file.