New name
Cleaned up calculation for error and tried to improve straightening. Now we mask around the first and last image of a batch to try to keep the scale more consistent across batches. This should result in fewer straightening seams, but there is still distortion if the camera is moving a lot.